* Support getting sample position from source stream after resync or initial load
