+++ NVorbis/Contracts/IPacket.cs
@@ -18,12 +18,7 @@
         /// <summary>
         /// Gets the granule position of the packet, if known.
         /// </summary>
-        long GranulePosition { get; }
-
-        /// <summary>
-        /// Gets the granule position of the framing page the packet is in.
-        /// </summary>
-        long PageGranulePosition { get; }
+        long? GranulePosition { get; }
 
         /// <summary>
         /// Gets whether the packet is the start of a new header set.
@@ -70,21 +65,6 @@
         /// <param name="count">The number of bits to read.</param>
         /// <returns>The value read.  If not enough bits remained, this will be a truncated value.</returns>
         ulong ReadBits(int count);
-
-        /// <summary>
-        /// Reads one bit from the packet and advances the read position.
-        /// </summary>
-        /// <returns><see langword="true"/> if the bit was a one, otehrwise <see langword="false"/>.</returns>
-        bool ReadBit();
-
-        /// <summary>
-        /// Reads into the specified buffer.
-        /// </summary>
-        /// <param name="buffer">The buffer to read into.</param>
-        /// <param name="index">The index into the buffer to use.</param>
-        /// <param name="count">The number of bytes to read into the buffer.</param>
-        /// <returns>The number of bytes actually read into the buffer.</returns>
-        int Read(byte[] buffer, int index, int count);
 
         /// <summary>
         /// Frees the buffers and caching for the packet instance.
